Android patch: Display names for countries/regions and currency symbol
1. Display names for Macedonia and Falkland Islands.

This change was originally introduced for English in Android by the
ICU52 upgrade:

https://android.googlesource.com/platform/external/icu/+/59d709d

As of the O release, this is replaced by consistently selecting the CLDR
variant name for all locales:

https://android.googlesource.com/platform/external/icu/+/ce738d3
https://android.googlesource.com/platform/external/icu/+/30d5dcd

It was later replaced by a patch in dataModifiers.txt in external/cldr.
https://r.android.com/692427

Bug: 36123938

2. Display name for Palestine.

This change was introduced in Android by the ICU4.4.1 upgrade:

https://android.googlesource.com/platform/external/icu/+/50294ea

It was later replaced by a patch in dataModifiers.txt in external/cldr.
https://r.android.com/692426

Bug: 8264703

3. Ukrainian hryvnia currency symbol.

Use the variant form instead of the hryvnia symbol:

https://unicode.org/cldr/trac/ticket/8814

This change was introduced in Android by this patch:

https://android.googlesource.com/platform/external/icu/+/cbdaa2b

It was later replaced by a patch in dataModifiers.txt in external/cldr.
https://r.android.com/692428

Bug: 21295835
Change-Id: Ie5b3d930e727d55752d7319cdbbeb15d3764b5eb