Expose libicuuc_headers to ExtServices
This is used to satisfy the dependency required by the library
tflite-support. The library only uses the MACRO U8_NEXT, so we just
need to take the header files.

ExtServices is a R mainline module and thus setting these fields:
    sdk_version: "current",
    min_sdk_version: "30",

Will move to the ndk header once mainline-prod is rebased to S.

Bug: 180325304

Test: m libicuuc_headers_for_app

Change-Id: Ifdaac239a99b4ee2be5d9e41d3d1dc7696a3c47d