Add ucnv_convert to libandroidicu
We already provide ucnv APIs in libandroidicu and thus there's no
harm adding ucnv_convert there. This API is requested in http://b/163081189

libandroidicu is not a stable API surface, but a stop-gap measure
to limit the API surface from ICU4C to the system/frameworks.

Note that libandroidicu doesn't provide the APIs for NDK, but libicu does.

Bug: 163081189
Test: m droid
Change-Id: Ice35b1d0242f9c721bd5e990a4c5d8cc1615033c