Add cmake files to external/icu
This is intended to allow building external/icu to be statically linked into standalone apps/libraries.

In particular, this will be compiled into AppSearch's JetPack library, which has a minSdkVersion of 14. It is not compiled against the Android NDK (though it could be!) nor is it expected to be used by 3rd-party Android app.

Bug: 183574903

Change-Id: Ie63044708cbe13b8e28fbad38a64114cf70d2323