Update the ICU binary file due to the latest change in aosp/master
Apparently, the "one" variant can be skipped when it's identical to
the "other" variant. However, I am not sure what caused the change
but it looks safe to do so.

The change may not be related to ICU 70 because the data has been the
same for many years, and the upstream .txt file has the "one" variant
in es_419.txt
https://github.com/unicode-org/icu/blob/release-70-1/icu4c/source/data/locales/es_419.txt#L14

Test: tools/updateicudata.py
Change-Id: I5ca89eded718bc835685982f28a8eb4fe98ec804