DO NOT MERGE Ignore TimeZoneTest#TestCanonicalID [O-MR1 CTS].
Test is part of CTS, so it should pass both on old builds (with 2021a3
and old tzdata) and 2022a. To achieve that subset of excluded2 list
needs to be commented out and non-obvious exemption list needs to be
introduced to check some common part of 2022a and 2021a3.

Bug: 228944103
Bug: 228807416

Test: atest CtsIcuTestCases:TimeZoneTest on ToT and ab/7679548 build
Change-Id: I5bb2fe5088725fa209177c7c87d01bc134866083
Merged-In: I5bb2fe5088725fa209177c7c87d01bc134866083