DO NOT MERGE Ignore tztest#TestCanonicalID [O-MR1 CTS].
Test is part of CTS, so it should pass both on old builds (with 2021a3
and old tzdata) and 2022a. To achieve that subset of excluded2 list
needs to be commented out and non-obvious exemption list needs to be
introduced to check some common part of 2022a and 2021a3.

Bug: 228944103
Bug: 227747309

Test: CtsIcu4cTestCases
Change-Id: Ia8dc0ccc03a1d4ad6565906c59e2f79986248c3d
Merged-In: Ia8dc0ccc03a1d4ad6565906c59e2f79986248c3d